rename: change name of directory of ValidDataException
