+++ src/utils/exception-global.filter.ts
@@ -5,7 +5,7 @@ import {
   HttpException,
   HttpStatus,
 } from '@nestjs/common';
-import { ValidDataException } from './validate-exceptions';
+import { ValidDataException } from './validate-data-exception';
 
 @Catch()
 export class ExceptionGlobalFilter<T> implements ExceptionFilter {
